+++ CHANGELOG.md
@@ -9,6 +9,9 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- Add support for [LingBot Vision](https://github.com/Robbyant/lingbot-vision) backbones
+  `dinov3/vits16-lingbot`, `dinov3/vitb16-lingbot`, and `dinov3/vitl16-lingbot`.
+
 ### Changed
 
 ### Deprecated